Reviewed-on: https://codefirst.iut.uca.fr/git/AllDev/Android/pulls/4

+++ src/data/src/main/java/fr/iut/alldev/allin/data/api/model/ApiUser.kt
@@ -25,7 +25,7 @@ data class ResponseUser(
         id = id,
         username = username,
         email = email,
-        coins = nbCoins.toInt()
+        coins = nbCoins
     )
 }
